refactoring: split controllers, add jackson configuration

+++ src/main/resources/db/changelog/init.sql
@@ -3,7 +3,7 @@
 --changeset Dmitry:1
 CREATE TABLE user_info
 (
-    user_id      SERIAL PRIMARY KEY,
+    user_id      BIGSERIAL PRIMARY KEY,
     phone_number VARCHAR(12)  NOT NULL,
     name         VARCHAR(100) NOT NULL,
     email        VARCHAR(100) NOT NULL,
@@ -13,8 +13,8 @@ CREATE TABLE user_info
 --changeset Dmitry:2
 CREATE TABLE record
 (
-    record_id   SERIAL PRIMARY KEY,
-    user_id     INT          NOT NULL REFERENCES user_info ON DELETE CASCADE,
+    record_id   BIGSERIAL PRIMARY KEY,
+    user_id     BIGINT          NOT NULL REFERENCES user_info ON DELETE CASCADE,
     key         VARCHAR(100) NOT NULL,
     value       TEXT         NOT NULL,
     revision    INT DEFAULT 0,
